pkg/trace/appsec: add the envoy auth grpc api stubs
