+++ pkg/trace/appsec/appsec.go
@@ -6,102 +6,13 @@
 package appsec
 
 import (
-	"encoding/json"
-	"io"
-	"net/http"
-	"time"
-
-	"github.com/DataDog/datadog-agent/pkg/trace/log"
-	waf "github.com/DataDog/go-libddwaf"
-	"github.com/pkg/errors"
+	"github.com/DataDog/go-libddwaf"
 )
 
-type (
-	requestPayload struct {
-		Type         string                 `json:"type"`
-		Trace        bool                   `json:"trace"`
-		ExtraTags    *spanTags              `json:"extra_tags"`
-		SecAddresses map[string]interface{} `json:"sec_addresses"`
-	}
-
-	spanTags struct {
-		Meta    map[string]string  `json:"meta"`
-		Metrics map[string]float64 `json:"metrics"`
-	}
-
-	responsePayload struct {
-		Type    string   `json:"type"`
-		Matches []byte   `json:"matches"`
-		Actions []string `json:"actions"`
-	}
-
-	errorPayload struct {
-		Type  string `json:"type"`
-		Error string `json:"error"`
-	}
-)
-
-const (
-	defaultWAFTimeout           = 4 * time.Millisecond
-	defaultObfuscatorKeyRegex   = `(?i)(?:p(?:ass)?w(?:or)?d|pass(?:_?phrase)?|secret|(?:api_?|private_?|public_?)key)|token|consumer_?(?:id|key|secret)|sign(?:ed|ature)|bearer|authorization`
-	defaultObfuscatorValueRegex = `(?i)(?:p(?:ass)?w(?:or)?d|pass(?:_?phrase)?|secret|(?:api_?|private_?|public_?|access_?|secret_?)key(?:_?id)?|token|consumer_?(?:id|key|secret)|sign(?:ed|ature)?|auth(?:entication|orization)?)(?:\s*=[^;]|"\s*:\s*"[^"]+")|bearer\s+[a-z0-9\._\-]+|token:[a-z0-9]{13}|gh[opsu]_[0-9a-zA-Z]{36}|ey[I-L][\w=-]+\.ey[I-L][\w=-]+(?:\.[\w.+\/=-]+)?|[\-]{5}BEGIN[a-z\s]+PRIVATE\sKEY[\-]{5}[^\-]+[\-]{5}END[a-z\s]+PRIVATE\sKEY|ssh-rsa\s*[a-z0-9\/\.+]{100,}`
-)
-
-func NewHTTPHandler() (http.Handler, error) {
-	handle, err := waf.NewHandle(staticRecommendedRules, defaultObfuscatorKeyRegex, defaultObfuscatorValueRegex)
+func NewWAFHandle() (*waf.Handle, error) {
+	wafHandle, err := waf.NewHandle(staticRecommendedRules, defaultObfuscatorKeyRegex, defaultObfuscatorValueRegex)
 	if err != nil {
-		return nil, errors.Wrap(err, "appsec: unexpected error while instantiating the waf")
+		return nil, err
 	}
-
-	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		// Parse the body
-		var reqPayload requestPayload
-		body := r.Body
-		defer func() {
-			io.ReadAll(body)
-			body.Close()
-		}()
-		if err := json.NewDecoder(body).Decode(&reqPayload); err != nil {
-			writeErrorResponse(w, errors.Wrap(err, "appsec: couldn't parse the request body into json:"))
-			return
-		}
-
-		wafCtx := waf.NewContext(handle)
-		if wafCtx == nil {
-			// The WAF handle got released in the meantime
-			writeUnavailableResponse(w)
-			return
-		}
-		defer wafCtx.Close()
-
-		matches, actions, err := wafCtx.Run(reqPayload.SecAddresses, defaultWAFTimeout)
-		if err != nil && err != waf.ErrTimeout {
-			writeErrorResponse(w, err)
-			return
-		}
-
-		if err := json.NewEncoder(w).Encode(responsePayload{
-			Type:    "waf_response",
-			Matches: matches,
-			Actions: actions,
-		}); err != nil {
-			log.Errorf("appsec: unexpected error while encoding the waf response payload into json: %v", err)
-		}
-	}), nil
-}
-
-func writeErrorResponse(w http.ResponseWriter, err error) {
-	w.WriteHeader(http.StatusInternalServerError)
-	log.Error(err)
-	res := errorPayload{
-		Type:  "error",
-		Error: err.Error(),
-	}
-	if err := json.NewEncoder(w).Encode(res); err != nil {
-		log.Errorf("appsec: couldn't encode the error response payload `%q` into json: %v", res, err)
-	}
-}
-
-func writeUnavailableResponse(w http.ResponseWriter) {
-	w.WriteHeader(http.StatusServiceUnavailable)
+	return wafHandle, nil
 }
